requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.10
MarkupSafe==2.1.3
msgpack-python==0.5.6
pymongo==4.4.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
